refactoring: 메서드를 세부 분리하고 클래스를 역할 별로 그룹화
