+++ src/main/java/menu/Application.java
@@ -4,7 +4,6 @@ import menu.system.MenuApplication;
 
 public class Application {
     public static void main(String[] args) {
-        // TODO: 프로그램 구현
         MenuApplication menuApplication = new MenuApplication();
         menuApplication.run();
     }
