+++ src/main/java/menu/controller/SetupController.java
@@ -1,5 +1,6 @@
 package menu.controller;
 
+import menu.controller.system.AbstractController;
 import menu.domain.Category;
 import menu.domain.MenuRepository;
 import menu.outputview.MainOutputView;
