+++ src/main/java/menu/controller/system/Controller.java
@@ -1,4 +1,4 @@
-package menu.controller;
+package menu.controller.system;
 
 public interface Controller {
     void process();
