+++ src/main/java/menu/controller/system/ControllerName.java
@@ -1,4 +1,4 @@
-package menu.controller;
+package menu.controller.system;
 
 public enum ControllerName {
     COACH_INFO, RECOMMANDING, SETUP
