+++ src/main/java/menu/inputview/AbstractInputView.java
@@ -2,7 +2,7 @@ package menu.inputview;
 
 import camp.nextstep.edu.missionutils.Console;
 
-public class AbstractInputView {
+public abstract class AbstractInputView {
     protected static String readInput() {
         return Console.readLine();
     }
