+++ src/main/java/menu/outputview/RecommendingOutputView.java
@@ -13,20 +13,31 @@ public class RecommendingOutputView {
 
     public static final String RECOMMENDING_RESULT_MESSAGE = "메뉴 추천 결과입니다.";
     public static final String RECOMMEND_COMPLETED_MESSAGE = "추천을 완료했습니다.";
+    public static final String CLOSING_BRACKET = " ]";
+    public static final String OPENING_BRACKET = "[ ";
+    public static final String VALUE_SEPARATOR = " | ";
+    public static final String DAYS_MESSAGE_PREFIX = "구분 | ";
+    public static final String CATEGORIES_MESSAGE_PREFIX = "카테고리 | ";
 
     public static void printRecommendingResult(List<String> categoryNames, Map<Coach, List<Menu>> recommendedMenus) {
         printResultMessage();
+        printRecommendedMenus(categoryNames, recommendedMenus);
+        printResultCompletedMessage();
+    }
 
+    private static void printRecommendedMenus(List<String> categoryNames, Map<Coach, List<Menu>> recommendedMenus) {
         printDays();
         printCategories(categoryNames);
         for (Map.Entry<Coach, List<Menu>> coachAndMenus : recommendedMenus.entrySet()) {
-            printOpening();
-            System.out.printf("%s | ", coachAndMenus.getKey().getName());
-            printMenuNames(coachAndMenus);
-            printClosing();
+            printRecommendedMenu(coachAndMenus);
         }
+    }
 
-        printResultCompletedMessage();
+    private static void printRecommendedMenu(Map.Entry<Coach, List<Menu>> coachAndMenus) {
+        printOpening();
+        System.out.printf("%s | ", coachAndMenus.getKey().getName());
+        printMenuNames(coachAndMenus);
+        printClosing();
     }
 
     private static void printResultCompletedMessage() {
@@ -51,7 +62,7 @@ public class RecommendingOutputView {
 
     private static void printCategories(List<String> categoryNames) {
         printOpening();
-        System.out.print("카테고리 | ");
+        System.out.print(CATEGORIES_MESSAGE_PREFIX);
         Iterator<String> iterator = categoryNames.iterator();
         printValues(iterator);
         printClosing();
@@ -59,25 +70,25 @@ public class RecommendingOutputView {
 
     private static void printDays() {
         printOpening();
-        System.out.print("구분 | ");
+        System.out.print(DAYS_MESSAGE_PREFIX);
         Iterator<String> iterator = Day.findAllNames().iterator();
         printValues(iterator);
         printClosing();
     }
 
     private static void printClosing() {
-        System.out.println(" ]");
+        System.out.println(CLOSING_BRACKET);
     }
 
     private static void printOpening() {
-        System.out.print("[ ");
+        System.out.print(OPENING_BRACKET);
     }
 
     private static void printValues(Iterator<String> iterator) {
         while (iterator.hasNext()) {
             System.out.print(iterator.next());
             if (iterator.hasNext()) {
-                System.out.print(" | ");
+                System.out.print(VALUE_SEPARATOR);
             }
         }
     }
